LayoutTests/media/modern-media-controls/macos-inline-media-controls/macos-inline-media-controls-compact-buttons-styles.html

<script src="../../../resources/js-test-pre.js"></script>
<script src="../resources/media-controls-utils.js" type="text/javascript"></script>
<script src="../resources/media-controls-loader.js" type="text/javascript"></script>
<body>
<script type="text/javascript">

description("Testing the <code>MacOSInlineMediaControls</code> computed styles in compact mode.");

window.jsTestIsAsync = true;

const mediaControls = new MacOSInlineMediaControls({ width: 240 });
mediaControls.layoutTraits = LayoutTraits.macOS | LayoutTraits.Compact;
document.body.appendChild(mediaControls.element);

const buttonPositions = [
    [mediaControls.playPauseButton, 6],
    [mediaControls.skipBackButton, 6],
    [mediaControls.muteButton, 6],
    [mediaControls.fullscreenButton, 6.5]
];

let button, y;
(function testButton(index)
{
    [button, y] = buttonPositions[index];
    debug("");
    debug(button.constructor.name);
    shouldBecomeEqualToString("getComputedStyle(button.element).webkitMaskPositionY", `${y}px`, () => {
        shouldBeEqualToString("getComputedStyle(button.element).height", "25px");
        if (index === buttonPositions.length - 1) {
            mediaControls.element.remove();
            finishJSTest();
        } else
            testButton(index + 1);
    });
})(0);

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
